+++ src/components/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Navigation from '@/components/Navigation';
+import { Navigation } from '@/components';
 
 const Layout = () => {
   return (
