+++ src/components/index.ts
@@ -1,0 +1,36 @@
+// Atoms
+export { default as Image } from './atoms/Image';
+export { default as Input } from './atoms/Input';
+export { default as Label } from './atoms/Label';
+export { default as Text } from './atoms/Text';
+export { default as TextArea } from './atoms/TextArea';
+
+// Molecules
+export { default as InputField } from './molcules/InputField';
+export { default as MessageInput } from './molcules/MesageInput';
+export { default as OrderCard } from './molcules/OrderCard';
+
+// Organisms
+export { default as CardCarousel } from './organisms/CardCarousel';
+export { default as ProductInfo } from './organisms/ProductInfo';
+export { default as ReceiverSection } from './organisms/ReceiverSection';
+export { default as SenderSection } from './organisms/SenderSection';
+
+//******Atomic Design Pattern적용 전 컴포넌트들도 추후에 Atomic Design Pattern적용 예정******//
+
+// Common
+export { default as Button } from './common/Button';
+export { default as IconButton } from './common/IconButton';
+export { default as ItemCard } from './common/ItemCard';
+
+// Form
+export { default as LoginInputField } from './Form/LoginInputField';
+
+// Layout
+export { default as Layout } from './Layout/Layout';
+
+// Navigation
+export { default as Navigation } from './Navigation';
+
+// Other
+export { default as ProtectedRoute } from './ProtectedRoute';
